requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
sqlalchemy
psycopg2-binary==2.9.10  # For PostgreSQL support
redis==4.6.0
rq==1.16.1
async-timeout>=4.0.2
click>=5
requests
beautifulsoup4
trafilatura
python-json-logger==2.0.7
email-validator==2.1.0.post1
bcrypt==4.1.2
PyJWT
openai>=1.0.0,<2.0.0
flask-restx>=1.1.0
apispec>=6.3.0
apispec-webframeworks>=0.5.2
marshmallow>=3.20.1
rq-dashboard
apify-client>=1.6.0
playwright
gunicorn==21.2.0
watchdog==6.0.0

# Log streaming system dependencies
websockets>=11.0.0
docker>=7.0.0
python-json-logger>=2.0.7

# --- Testing dependencies ---
pytest==8.0.2
pytest-cov==4.1.0
pytest-watch==4.2.0